Update NgGrid.ts
Fixes dragging and resizing issues when something is selected inside the grid item.

Please REVIEW! I am unable to test document.selection at this moment, so please double-check before merging to your repository.
Consider moving it to your dragStart and resizeStart functions, as I purely upload a hotfix that works. 

It might be more resource-efficient to move it to the xStart functions, but remember that if you drag the NgGridItem it might select the parts inside, and breaking it again while moving.

Therefore, we need testing to make sure this is as optimized as it can be.